Merge Var and Func
This is a slight variation of `chibicc`. rui merges Obj and Func
together and uses a boolean `is_func` to differentiate between them
and then uses a `is_local` to differentiate local and global vars. I
don't like that approach because I think it will be too buggy. So I use
a struct that has a kind that can be either `Function` or `Variable`
(possibly others like struct, typedef etc in the future) and a union
that has this payload. This should make it more easier to manage than
`is_local` , `is_function` etc